+++ types.nim
@@ -1,6 +1,6 @@
 type
   # Any generic Vector is assumed to be a column vector.
-  Vector* = seq[int]
+  Vector* = seq[float]
   RowVector* = Vector # can this be distinct?
   Matrix* = seq[RowVector]
